working: add more hash api's fix some which weren't working as expected

+++ src/logical_allocator.h
@@ -30,7 +30,7 @@
 #include "hash_arena.h"
 #include "page_modifications.h"
 
-typedef uint16_t PageSizeType;
+typedef uint32_t PageSizeType;
 typedef heap::set<size_t> address_set;
 enum {
     LPageSize = page_size - sizeof(storage)
